+++ packages/reactivity/src/dependency.js
@@ -1,8 +1,9 @@
 import { Scheduler } from './scheduler.js';
 
 export class Dependency {
-  constructor() {
+  constructor(...metadata) {
     this.subscribers = new Set();
+    this.setContext(metadata);
   }
 
   depend() {
@@ -12,14 +13,27 @@ export class Dependency {
     }
   }
 
-  changed(context) {
-    this.subscribers.forEach(subscriber => subscriber.invalidate(context));
+  // allows metadata to be passed with dependency for debugging
+  setContext(context = {}) {
+    if (Error.captureStackTrace) {
+      Error.captureStackTrace(context, this.setContext);
+    } else {
+      context.stack = new Error().stack;
+    }
+    this.context = context;
   }
 
+  changed(context) {
+    this.setContext(context);
+    this.subscribers.forEach(subscriber => subscriber.invalidate(this.context));
+  }
+
+  // called after flush
   cleanUp(reaction) {
     this.subscribers.delete(reaction);
   }
 
+  // identical for now but called from stop()
   unsubscribe(reaction) {
     this.subscribers.delete(reaction);
   }